requirements.txt
asgiref==3.10.0
boto3==1.40.59
botocore==1.40.59
Django==5.2.7
djangorestframework==3.16.1
jmespath==1.0.1
python-dateutil==2.9.0.post0
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
